totally reorganized the entire qombat section
"Summary" of combat at beginning almost explained everything in one go...
so now it actually does.
then the details are separated into sections below.

I threw in a formula for grazing damage with very little thought,
so we'll have to think about that.

A graze is now zero, hit positive, miss negative.
This better matches how general opposed actions work.
(Assuming we are qul with them working this way)

I removed notes that we've addressed and shuffled the blob of notes into better places.